Added a ServersPage container componenet and moved Servers into presentational
